finally landed on weights
